Breaking down into multiple smaller files and finishing organization
